# optimize node startup speed and memory allocation