--- a/ethereum/api/src/integration-test/java/org/hyperledger/besu/ethereum/api/jsonrpc/BlockchainImporter.java
+++ b/ethereum/api/src/integration-test/java/org/hyperledger/besu/ethereum/api/jsonrpc/BlockchainImporter.java
@@ -44,7 +44,7 @@ public class BlockchainImporter {
   public BlockchainImporter(final URL blocksUrl, final String genesisJson) throws Exception {
     protocolSchedule =
         MainnetProtocolSchedule.fromConfig(
-            GenesisConfigFile.fromConfig(genesisJson).getConfigOptions(),
+            GenesisConfigFile.fromConfigWithoutAccounts(genesisJson).getConfigOptions(),
             MiningParameters.newDefault(),
             new BadBlockManager());
     final BlockHeaderFunctions blockHeaderFunctions =