@ahamlat 2024-04-16
@lyfsn Great work on the PR, I have few questions if you don't mind :

- Is there a way to reproduce the numbers on my side, by sharing the big (fake) genesis file, and the configuration you used for your test ? 
- Is "requiring less than 2GB of memory" a strong prerequisite for your use case ? 
- Can you share the -Xmx value used during your test?
- Have you checked GC activity during the test period ? 
- Have you done some profiling to see what is taking more than 7 minutes in the first and third test with this PR's code ?