@lyfsn 2024-04-16
> @lyfsn Great work on the PR, I have few questions if you don't mind :
> 
> * Is there a way to reproduce the numbers on my side, by sharing the big (fake) genesis file, and the configuration you used for your test ?
> * Is "requiring less than 2GB of memory" a strong prerequisite for your use case ?
> * Can you share the -Xmx value used during your test?
> * Have you checked GC activity during the test period ?
> * Have you done some profiling to see what is taking more than 7 minutes in the first and third test with this PR's code ?

- You can use genesis.json from this repository (though it's smaller than the one I used). [LInk](https://github.com/OpenFusionist/network_config)
- This repository has 8 `jfr` files; they are about memory allocation records. [LInk](https://github.com/lyfsn/besu-starter?tab=readme-ov-file#jfr)
- Also, in this repository, there are some records about profiler records. [Link](https://github.com/lyfsn/besu-starter?tab=readme-ov-file#empty-data-directory-on-startup-startup-time-7-minutes)
- I run the binary Besu directly, so I do not set -Xmx explicitly. I run this on a 64GB server, so the default should be 16GB.

You can refer to this information first, and if you have more specific needs, please feel free to let me explain!