@ahamlat 2024-04-18
@lyfsn thanks for the troubleshooting data, awesome work.
I was able to reproduce some of the number with this PR, for `--genesis-state-hash-cache-enabled=true && block number=0`, but then I did the test on a previous release, 23.10.2 (without both PRs), and it was faster for the block 0 test :

- `block number=0` : 5 minutes
- `block number=1` : 4 minutes 17 seconds 

I'm currently analyzing the regression as we know what code is causing it.
I would like to know what was the test you shared in you blog where Besu took more than 100 minutes to start