@lyfsn 2024-04-18
> @lyfsn thanks for the troubleshooting data, awesome work. I was able to reproduce some of the number with this PR, for `--genesis-state-hash-cache-enabled=true && block number=0`, but then I did the test on a previous release, 23.10.2 (without both PRs), and it was faster for the block 0 test :
> 
> * `block number=0` : 5 minutes
> * `block number=1` : 4 minutes 17 seconds
> 
> I'm currently analyzing the regression as we know what code is causing it. I would like to know what was the test you shared in you blog where Besu took more than 100 minutes to start

I guess there are two main reasons:
- At that time, we used our devnet file, which is about 1.1GB, a bit larger than the mainnet file, which is about 800MB. The mainnet file was provided to you for convenience in using publicly available files for testing.
- The server's hardware was limited at that time, with only 16GB of RAM and a set 16GB swap space. The swap space obviously slowed down the processing speed.

These were the parameters for the test at that time:
|Client|Startup command|Version|OS|Arch|CPU core|RAM|Swap space|
|--|--|--|--|--|--|--|--|
|Besu|env BESU_OPTS=-Xmx16g ./besu --(....)|24.3.0/openjdk17|ubuntu 22.04|x64|2|16GB|16GB|

Additionally, are you referring to this blog article? The article also detailed the testing conditions:
https://devlog.fusionist.io/posts/ethereum-execution-clients-memory-comparison-in-endurance/