@fab-10 2024-04-18
are your genesis files public? I would like to test them on the refactor, if not I will try to generate one with random data