@fab-10 2024-04-22
@lyfsn could you test [this PR](https://github.com/hyperledger/besu/pull/6977)? 
The allocations are lazy loaded only when needed, and there is reduced memory usage when parsing it, with less object copies and allocations streaming.